requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-yasg==1.21.7
sqlparse==0.5.1
tzdata==2024.1
psycopg2==2.9.9
python-dotenv==1.0.1
inflection==0.5.1
packaging==24.1
pytz==2024.1
pyyaml==6.0.2
uritemplate==4.1.1
setuptools==74.0.0
redis==5.0.8
amqp==5.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
kombu==5.4.0
prompt-toolkit==3.0.47
python-dateutil==2.9.0.post0
six==1.16.0
vine==5.1.0
wcwidth==0.2.13
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
cron-descriptor==1.4.5
django-celery-beat==2.7.0
django-timezone-field==7.0
python-crontab==3.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
requests==2.32.3
urllib3==2.2.2

